Refactor: Rename MainMenu to MainMenuOverlay
